requirements.txt
torch
pyyaml